+++ src/appHidden.ts
@@ -1,0 +1,14 @@
+import { app } from 'electron';
+
+let appHidden = false;
+
+export const setAppHidden = (hidden: boolean) => {
+  appHidden = hidden;
+  if (hidden) {
+    app?.hide();
+  }
+};
+
+export const getAppHidden = () => {
+  return appHidden;
+};