+++ src/background.ts
@@ -1,26 +1,13 @@
 import { grep } from 'shelljs';
 import log from 'electron-log';
-import { ChildProcess, fork } from 'child_process';
-import {
-  KIT,
-  KENV,
-  execPath,
-  NODE_PATH,
-  PATH,
-  DOTENV_CONFIG_PATH,
-  KIT_MAC_APP,
-} from './helpers';
-import { getVersion } from './version';
-import { consoleLog, getLog } from './logs';
+
+import { createMessageHandler } from './messages';
+import { TOGGLE_BACKGROUND } from './channels';
+import { emitter } from './events';
+import { backgroundMap, Background } from './state';
+import { createChild } from './run';
 
 const backgroundMarker = 'Background: ';
-
-interface Background {
-  child: ChildProcess;
-  start: string;
-}
-
-export const backgroundMap = new Map<string, Background>();
 
 export const removeBackground = (filePath: string) => {
   if (backgroundMap.get(filePath)) {
@@ -32,48 +19,21 @@ export const removeBackground = (filePath: string) => {
   }
 };
 
-export const backgroundScript = (filePath: string, runArgs: string[]) => {
-  const child = fork(filePath, [...runArgs], {
-    silent: true,
-    // stdio: 'inherit',
-    execPath,
-    execArgv: ['--require', 'dotenv/config', '--require', KIT_MAC_APP],
-    env: {
-      ...process.env,
-      KIT_CONTEXT: 'app',
-      KIT_MAIN: filePath,
-      PATH,
-      KENV,
-      KIT,
-      NODE_PATH,
-      DOTENV_CONFIG_PATH,
-      KIT_APP_VERSION: getVersion(),
-    },
+export const backgroundScript = (scriptPath: string, runArgs: string[]) => {
+  const child = createChild({
+    from: 'background',
+    scriptPath,
+    runArgs,
   });
 
   const pid = child?.pid;
   child?.on('exit', () => {
-    if (backgroundMap.get(filePath)?.child?.pid === pid) {
-      log.info(`> exit background process: ${filePath} ${pid}`);
-      backgroundMap.delete(filePath);
+    if (backgroundMap.get(scriptPath)?.child?.pid === pid) {
+      backgroundMap.delete(scriptPath);
     }
   });
 
-  child?.on('message', (data: any) => {
-    switch (data?.channel) {
-      case 'CONSOLE_LOG':
-        console.log(data);
-        getLog(data?.kitScript).info(data.log);
-        break;
-      case 'CONSOLE_WARN':
-        getLog(data?.kitScript).warn(data.log);
-        break;
-      default:
-        consoleLog.info(`background: Unknown message ${data.channel}`);
-    }
-  });
-
-  log.info(`> start background process: ${filePath} ${child.pid}`);
+  child?.on('message', createMessageHandler('background'));
 
   return child;
 };
@@ -111,23 +71,6 @@ export const updateBackground = (filePath: string, fileChange = false) => {
   }
 };
 
-export const getBackgroundTasks = () => {
-  const tasks = Array.from(backgroundMap.entries()).map(
-    ([filePath, { child, start }]: [string, Background]) => {
-      return {
-        filePath,
-        process: {
-          spawnargs: child?.spawnargs,
-          pid: child?.pid,
-          start,
-        },
-      };
-    }
-  );
-
-  return tasks;
-};
-
 export const toggleBackground = (filePath: string) => {
   if (backgroundMap.get(filePath)) {
     removeBackground(filePath);
@@ -135,3 +78,7 @@ export const toggleBackground = (filePath: string) => {
     updateBackground(filePath);
   }
 };
+
+emitter.on(TOGGLE_BACKGROUND, (data) => {
+  toggleBackground(data.filePath as string);
+});