+++ src/logs.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 /* eslint-disable import/prefer-default-export */
 import log from 'electron-log';
 import { kenvPath } from './helpers';
@@ -7,14 +8,25 @@ consoleLog.transports.file.resolvePath = () => kenvPath('logs', 'console.log');
 
 const logMap = new Map<string, log.ElectronLog>();
 export const getLog = (id: string) => {
-  const normalizedId = id.replace(/.*\//g, '');
+  try {
+    const normalizedId = id.replace(/.*\//g, '').replace('.js', '');
 
-  if (logMap.get(normalizedId))
-    return logMap.get(normalizedId) as log.ElectronLog;
+    if (logMap.get(normalizedId))
+      return logMap.get(normalizedId) as log.ElectronLog;
 
-  const scriptLog = log.create(normalizedId);
-  scriptLog.transports.file.resolvePath = () =>
-    kenvPath('logs', `${normalizedId}.log`);
+    const scriptLog = log.create(normalizedId);
+    scriptLog.transports.file.resolvePath = () =>
+      kenvPath('logs', `${normalizedId}.log`);
 
-  return scriptLog;
+    return scriptLog;
+  } catch {
+    return {
+      info: (...args: any[]) => {
+        console.log(...args);
+      },
+      warn: (...args: any[]) => {
+        console.warn(...args);
+      },
+    };
+  }
 };